[LIGO-510] Gist project sharing
Problem: as now we are using local projects instead of remote we
need an ability to share them.

Solution: we are using gist to share the code. We use file paths
as file names a simply create new gist to share the project. And
when we want to import it we use file names as paths again.