[LIGO-577] Restructure
Problem: current structure of the Ligo IDE is more
like monorepo with a lot of packages. It was a good
solution for EthereumStudio as for the modular IDE.
But for our solution we do not need it at all as it
cause different problems on development stage.

Solution: remove such monorepo structure and put all
app code to the single package.